:recycle: chore(editor): Extract registering file associations to a lib
