+++ lib/either/src/operators/index.ts
@@ -2,6 +2,7 @@ export * from "./bichain"
 export * from "./bimap"
 export * from "./bitap"
 export * from "./chain"
+export * from "./fix"
 export * from "./left-chain"
 export * from "./left-map"
 export * from "./left-tap"
