+++ lib/frontend-stream-file-associations/readme.md
@@ -1,0 +1,1 @@
+# Frontend Stream File Associations
